fix: harden connector payload and lookup paths
