+++ packages/connector/src/client/AGENTS.md
@@ -31,6 +31,7 @@
   - `outbound-flush.ts` (queue flush orchestration and serialization assistance so `ConnectorClient` no longer reaches directly into `queue.ts`).
   - `inbound-router.ts` (handles raw message parsing, routes heartbeat/deliver frames to heartbeat manager/handlers, and records metrics before handing off to `handleInboundDeliverFrame`).
 - `delivery.ts` and `inbound-delivery.ts` stay responsible for OpenClaw delivery + ack orchestration and should expose injectable hooks for testing retries/timeout logic.
+- Keep hook payload shaping itself in the shared `../openclaw-payload.ts` helper instead of copy-pasting payload JSON builders across client/runtime delivery paths.
 - `delivery.ts` may resolve optional sender profile metadata via injected resolver callbacks, but delivery must stay best-effort: lookup failures cannot block ACK flow.
 - When sender profile is unavailable, omit `x-clawdentity-agent-name` and `x-clawdentity-display-name` instead of sending empty placeholders.
 - `/hooks/wake` payload shaping in `delivery.ts` must preserve inbound `sessionId` when present so explicit OpenClaw thread routing is not lost.
